requirements.txt
streamlit
langchain
faiss-cpu
transformers
datasets
torch
accelerate
sentence-transformers
scikit-learn
pandas
numpy
matplotlib
python-dotenv
tqdm
requests
aiohttp
